requirements.txt
langchain_google_genai
google-generativeai
streamlit
langchain
python-dotenv